+++ include/receivedframe.h
@@ -165,7 +165,7 @@ public:
      *  Total size of the frame (headers + payload)
      *  @return uint32_t
      */
-    uint64_t totalSize() const
+    size_t totalSize() const
     {
         // payload size + size of headers and end of frame byte
         return _payloadSize + 8;
